+++ tests/test_action_expander.py
@@ -1,26 +1,8 @@
-import re
-
 import pytest
 
 from pycfmodel import parse
-from pycfmodel.action_expander import _build_regex, _expand_action, _expand_actions
+from pycfmodel.action_expander import _expand_action, _expand_actions
 from pycfmodel.cloudformation_actions import CLOUDFORMATION_ACTIONS
-
-
-@pytest.mark.parametrize(
-    "action, expected_pattern",
-    [
-        ("", re.compile("^$", re.IGNORECASE)),
-        ("*", re.compile("^.*$", re.IGNORECASE)),
-        ("?", re.compile("^.{1}$", re.IGNORECASE)),
-        ("a", re.compile("^a$", re.IGNORECASE)),
-        ("a*", re.compile("^a.*$", re.IGNORECASE)),
-        ("a?", re.compile("^a.{1}$", re.IGNORECASE)),
-        ("a*a?", re.compile("^a.*a.{1}$", re.IGNORECASE)),
-    ],
-)
-def test_build_regex(action, expected_pattern):
-    assert _build_regex(action) == expected_pattern
 
 
 @pytest.mark.parametrize(